+++ Basil/include/Platform/OpenGL/OpenGLFramebuffer.h
@@ -17,6 +17,7 @@ namespace Basil
 			void unbind() override;
 			void resize(uint32_t width, uint32_t height) override;
 			int readPixel(uint32_t attachmentIndex, int x, int y) override;
+			void clearAttachment(uint32_t attachmentIndex, int value) override;
 			uint32_t getColorAttachmentRendererID(uint32_t index = 0) const override;
 			const FramebufferSpecification& getSpecification() const override;
 		private: