+++ Basil/include/Renderer/Framebuffer.h
@@ -61,6 +61,7 @@ namespace Basil
 			virtual void unbind() = 0;
 			virtual void resize(uint32_t width, uint32_t height) = 0;
 			virtual int readPixel(uint32_t attachmentIndex, int x, int y) = 0;
+			virtual void clearAttachment(uint32_t attachmentIndex, int value) = 0;
 			virtual uint32_t getColorAttachmentRendererID(uint32_t index = 0) const = 0;
 			virtual const FramebufferSpecification& getSpecification() const = 0;
 			static Shared<Framebuffer> create(const FramebufferSpecification& spec);