+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 Change log for the Basil Game Engine project.
 
+## 2021-05-17
+### Added
+* Added API to clear texture attachments.
+
+
 ## 2021-05-15
 ### Added
 * GL_RED_INTEGER as a framebuffer format.